backend/pyproject.toml
[project]
name = "otodb"
version = "0.1.0"
description = "otoDB is a community-driven website consisting of a collaborative user-managed database and wiki."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "diff-match-patch>=20241021",
    "django-comments-xtd>=2.10.1",
    "django-cors-headers>=4.6.0",
    "django-model-utils>=5.0.0",
    "django-simple-history>=3.7.0",
    "django-tagulous>=2.1.0",
    "django>=5.1.4",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "yt-dlp>=2025.09.23",
    "django-ninja>=1.3.0",
    "requests>=2.32.3",
    "django-markdownfield>=0.11.0",
    "django-debug-toolbar>=5.0.1",
    "django-ordered-model>=3.7.4",
    "granian>=2.1.0",
    "psycopg[binary,pool]>=3.2.6",
    "nh3>=0.2.21",
    "furl>=2.1.4",
    "django-cte>=2.0.0",
    "boto3>=1.40.30",
]

[tool.ruff.lint.per-file-ignores]
# Imports needed for Django to function
"otodb/models/**/__init__.py" = ["F401"]
